tests: Move self tests to root directory
