+++ setup.cfg
@@ -1,5 +1,5 @@
 [nosetests]
-tests=lisa/tests/lisa
+tests=tests/
 # The default regex can match things like create_local_testenv
 # We don't want that
 match=^[Tt]est_
