+++ tests/traces/plat_info.yml
