+++ tests/traces/sched_load/plat_info.yml
