+++ tests/traces/sched_load_avg/plat_info.yml
